requirements.txt
asgiref==3.2.3
Django==2.2.8
django-tinymce==2.8.0
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
